File changed on disk: reload or cancel
When we load or save a file, we save also the modified time. When the
document view receives the focus, we check if there are external changes
(if the modified time is more recent). A warning message (infobar) is
showed and the user can choose to reload the file or cancel.