Editor preferences using GSettings
GSettings is a lot easier than a key-value file, and gconf will be
deprecated with Gnome 3.0, so I decided to use GSettings.

That means that LaTeXila 2.0 will be released only after the new release
of GLib (2.26), in October.

That also means that Vala 0.9.2 minimum is required, as well as GLib
2.25.10 (unstable release).

The preferences dialog is made with Glade.

The editor preferences:
	- display line numbers
	- tab width
	- insert spaces instead of tabs
	- highlight current line
	- highlight matching brackets