Editor preferences: font
The user can choose to use the system fixed width font, or choose an
other font.

For the moment, the system font is not taken from gsettings but it
returns always the same font (Monospace 10). It's because I don't use an
unstable release of Gnome 3 for the moment. But the code (exactly 3 lines)
is commented. Therefore, as soon as I use an unstable version of
Gnome 3, the code will be uncommented and the real default system font
will be used.