Documents: move a document to a new window
We recreate a new tab in a new window with the same view. By doing this
we avoid to reload the file, and we also avoid headache with signals
(disconnect from the old window and reconnect in the new window).

If one day we want to support drag and drop of tabs between windows,
then we must keep the same tab, but this feature will not be
implemented in the near future.