Bug fixed: loss of data when saving
The fix was very easy: replace text.length (which is the number of
characters) by text.size() (which is the number of bytes). You can see
easily why some data could be lost before, because a character can be
more than one byte long.

This was a confusion with the API in which the method paramater have
the name "length".