Goto Line
Now the Goto Line is a HBox placed below the text view. In the future I
want to place it at the top right corner of the text view, like in
MonoDevelop 2.4. In Gedit that's a popup window, but I prefer the
behaviour of MonoDevelop.

Why not doing like MonoDevelop right now? I've tried to do it, but it
didn't work, it's more difficult as expected. For example it
envolves to create a custom container. There are more important stuff to
do, so this work will be done later.