Search: almost finished
It's incremental after 3 characters. Below that it's too slow with long
documents.

All the matching words are highlighted. There are some information in
the find entry ("X of N", or "N matches", or "Not found", ...).
It's interactive: if we delete or insert some text, the search is
updated (and there were a lot of specials cases to manage...).

What is missing:
	- the entire word only option
	- TextTag according to the style scheme